Refactor LoginPage tests
